fix(compiler): fixing husky to run prettier and lint on commit - using double quote for Sol; - increase max state count to 21; -

+++ contracts/interfaces/IERC20Extended.sol
@@ -3,33 +3,33 @@ pragma solidity 0.6.12;
 // With aditional optional views
 
 interface IERC20Extended {
-    // ERC20 Optional Views
-    function name() external view returns (string memory);
+  // ERC20 Optional Views
+  function name() external view returns (string memory);
 
-    function symbol() external view returns (string memory);
+  function symbol() external view returns (string memory);
 
-    function decimals() external view returns (uint8);
+  function decimals() external view returns (uint8);
 
-    // Views
-    function totalSupply() external view returns (uint);
+  // Views
+  function totalSupply() external view returns (uint256);
 
-    function balanceOf(address owner) external view returns (uint);
+  function balanceOf(address owner) external view returns (uint256);
 
-    function allowance(address owner, address spender) external view returns (uint);
+  function allowance(address owner, address spender) external view returns (uint256);
 
-    // Mutative functions
-    function transfer(address to, uint value) external returns (bool);
+  // Mutative functions
+  function transfer(address to, uint256 value) external returns (bool);
 
-    function approve(address spender, uint value) external returns (bool);
+  function approve(address spender, uint256 value) external returns (bool);
 
-    function transferFrom(
-        address from,
-        address to,
-        uint value
-    ) external returns (bool);
+  function transferFrom(
+    address from,
+    address to,
+    uint256 value
+  ) external returns (bool);
 
-    // Events
-    event Transfer(address indexed from, address indexed to, uint value);
+  // Events
+  event Transfer(address indexed from, address indexed to, uint256 value);
 
-    event Approval(address indexed owner, address indexed spender, uint value);
+  event Approval(address indexed owner, address indexed spender, uint256 value);
 }
