+++ contracts/test/IWETH.sol
@@ -1,7 +1,9 @@
 pragma solidity 0.6.12;
 
 interface IWETH {
-    function deposit() external payable;
-    function approve(address guy, uint wad) external returns (bool);
-    function transfer(address dst, uint wad) external returns (bool);
+  function deposit() external payable;
+
+  function approve(address guy, uint256 wad) external returns (bool);
+
+  function transfer(address dst, uint256 wad) external returns (bool);
 }
